app-query-book/requirements.txt
pika==1.0.0
mysql-connector-python==8.0.21